// src/modules/samity/interfaces/samity.interface.ts
export interface ISamityAttrs {
  id?: number;
  samityCode?: number;
  samityName?: string;
  doptorId?: number;
  officeId?: number;
  projectId?: number;
  districtId?: number;
  cityCorpId?: number;
  wardId?: number;
  upazilaId?: number;
  unionId?: number;
  vilageName?: string;
  address?: string;
  workPlaceLat?: string;
  workPlaceLong?: string;
  workingAreaRadius?: number;
  memberMinAge?: number;
  memberMaxAge?: number;
  samityMinMember?: number;
  samityMaxMember?: number;
  groupMinMember?: number;
  groupMaxMember?: number;
  branchCode?: number;
  foCode?: number;
  coopStatus?: boolean;
  coopRegNumber?: string;
  shareAmount?: number;
  weeklyMeetingDay?: string;
  authorizeStatus?: string;
  authorizedBy?: string;
  authorizedAt?: Date;
  createdBy?: string;
  createdAt?: Date;
  updatedBy?: string;
  updatedAt?: Date;
  upaCityId?: number;
  members?: any[];
}